+++ CHANGELOG.md
@@ -1,6 +1,22 @@
 ## Change log
 ----------------------
 
+Version 1.12
+-------------
+
+ADDED:
+
+- new enum for base rest path values
+- new enum for base action that follows after base rest paths
+
+CHANGED:
+
+- update of gradle plugin dependency 'com.diffplug.spotless:spotless-plugin-gradle' to new version 6.3.0
+- update of gradle plugin dependency 'io.freefair.gradle:lombok-plugin' to new version 6.4.1
+- update of dependency spring to new version 5.3.16
+- update of dependency spring-data to new version 2.6.2
+- update of spring-batch dependency to version 4.3.5
+
 Version 1.11
 -------------